Changes after review of pull request
