+++ examples/dataflow-scala-kafka2avro/.gitignore
@@ -4,3 +4,4 @@
 target
 .ensime*
 *~
+.idea
