+++ examples/dataflow-scala-kafka2avro/project/build.properties
@@ -1,1 +1,1 @@
-sbt.version=1.2.6
+sbt.version=1.2.8
